requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# by_origin : python train_simple.py --img 640 --batch-size 16 --epochs 20 --data data\kaist-rgbt.yaml --cfg models\yolov5n_kaist-rgbt.yaml --weights yolov5n.pt --workers 16 --name yolov5n-rgbt --entity codingosu-hanyang-university --rgbt --single-cls
# by : python train_simple.py --img 640 --batch-size 16 --epochs 10 --data data\kaist-rgbt.yaml --cfg models\yolov5n_kaist-rgbt.yaml --weights yolov5n.pt --workers 6 --name yolov5n-rgbt --entity codingosu-hanyang-university --rgbt --single-cls
# python train_simple.py --img 640 --batch-size 16 --epochs 10 --data data\kaist-rgbt.yaml --cfg models\yolov5n_kaist-rgbt.yaml --weights yolov5n.pt --workers 4 --name yolov5n-rgbt --entity codingosu-hanyang-university --rgbt --single-cls
# python train_simple.py --img 640 --batch-size 64 --epochs 10 --data data\kaist-rgbt.yaml --cfg models\yolov5n_kaist-rgbt.yaml --weights yolov5n.pt --workers 4 --name yolov5n-rgbt --entity codingosu-hanyang-university --rgbt --single-cls
# python train_simple.py --img 640 --batch-size 16 --epochs 1 --data data\kaist-rgbt.yaml --cfg models\yolov5n_kaist-rgbt.yaml --weights yolov5n.pt --workers 6 --name yolov5n-rgbt --entity codingosu-hanyang-university --rgbt --single-cls

# python train_simple_fold.py --img 640 --batch-size 16 --epochs 1 --data data\kaist-rgbt.yaml --cfg models\yolov5n_kaist-rgbt.yaml --weights yolov5n.pt --workers 6 --name yolov5n-rgbt --entity codingosu-hanyang-university --rgbt --single-cls --folds 2

# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib==3.10.1
numpy==1.23.5
opencv-python==4.11.0.86
pillow==11.2.1
psutil==7.0.0  # system resources
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
thop>=0.1.1  # FLOPs computation
# by 주석 torch==2.7.0  # see https://pytorch.org/get-started/locally (recommended)
# by 주석 torchvision==0.22.0
tqdm==4.67.1
ultralytics==8.3.128
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging ---------------------------------------------------------------------
# tensorboard>=2.4.1
# clearml>=1.2.0
# comet
wandb==0.17.0

# Plotting --------------------------------------------------------------------
pandas==2.2.3
seaborn==0.13.2

# Export ----------------------------------------------------------------------
# coremltools>=6.0  # CoreML export
# onnx>=1.10.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn<=1.1.2  # CoreML quantization
# tensorflow>=2.4.0,<=2.13.1  # TF exports (-cpu, -aarch64, -macos)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev>=2023.0  # OpenVINO export

# Deploy ----------------------------------------------------------------------
setuptools==80.3.1 # Snyk vulnerability fix
# tritonclient[all]~=2.24.0

# Extras ----------------------------------------------------------------------
# ipython  # interactive notebook
# mss  # screenshots
# albumentations>=1.0.3
pycocotools==2.0.7  # COCO mAP
wheel==0.45.1 # not directly required, pinned by Snyk to avoid a vulnerability
ipdb==0.13.13